Refactor with classes, rename helpers to utils
